refactor(plugins-api): Separate plugin analysis from code generation
Introduce an intermediary `PluginSpec` model that is used as input for
the code generation and move the code to create a `PluginSpec` to the
new class `PluginSpecFactory`.

This provides a better separation of concerns and makes it easier to
generate other output files based on the `PluginSpec`.
