+++ plugins/api/build.gradle.kts
@@ -27,5 +27,6 @@ dependencies {
 
     implementation(libs.kotlinpoet)
     implementation(libs.kotlinpoet.ksp)
+    implementation(libs.kotlinx.serialization.json)
     implementation(libs.ksp)
 }
